requirement.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==6.62.1
urllib3==1.26.6
